Rename AugmentCategory to AugmentCapsule
