+++ src/augmenting/AugmentCapsuleDisplay/SingleAugmentLine.tsx
